+++ TouchUpCore/TUCCursorUtilities.m
@@ -109,8 +109,8 @@
         self.cursorClickCount = 1;
     }
     
-    else if ((aLocation.x - self.locationOfLastClick.x) > self.doubleClickTolerance
-             && (aLocation.y - self.locationOfLastClick.y) > self.doubleClickTolerance) {
+    else if (fabs(aLocation.x - self.locationOfLastClick.x) > self.doubleClickTolerance
+             || fabs(aLocation.y - self.locationOfLastClick.y) > self.doubleClickTolerance) {
         // touch is too far away
         self.cursorClickCount = 1;
     }